Create TestCaseMixin so that TestCase and TransactionTestCase have access to the same helper functions.
