+++ nautobot/utilities/testing/__init__.py
@@ -5,7 +5,7 @@ from contextlib import contextmanager
 from celery.contrib.testing.worker import start_worker
 from django.apps import apps
 from django.contrib.auth import get_user_model
-from django.test import tag, TransactionTestCase as _TransactionTestCase
+from django.test import Client, tag, TransactionTestCase as _TransactionTestCase
 
 from nautobot.core.celery import app
 from nautobot.extras.context_managers import web_request_context
@@ -13,6 +13,7 @@ from nautobot.extras.jobs import run_job
 from nautobot.extras.management import populate_status_choices
 from nautobot.extras.models import JobResult
 from nautobot.extras.utils import get_job_content_type
+from nautobot.utilities.testing.mixins import TestCaseMixin
 
 from .api import APITestCase, APIViewTestCases
 from .utils import (
@@ -44,6 +45,9 @@ __all__ = (
     "run_job_for_testing",
 )
 
+# Use the proper swappable User model
+User = get_user_model()
+
 
 def run_job_for_testing(job, data=None, commit=True, username="test-user", request=None):
     """Provide a common interface to run Nautobot jobs as part of unit tests.
@@ -71,7 +75,6 @@ def run_job_for_testing(job, data=None, commit=True, username="test-user", reque
     if request and request.user:
         user_instance = request.user
     else:
-        User = get_user_model()
         user_instance, _ = User.objects.get_or_create(
             username=username, defaults={"is_superuser": True, "password": "password"}
         )
@@ -96,7 +99,7 @@ def run_job_for_testing(job, data=None, commit=True, username="test-user", reque
 
 
 @tag("unit")
-class TransactionTestCase(_TransactionTestCase):
+class TransactionTestCase(_TransactionTestCase, TestCaseMixin):
     """
     Base test case class using the TransactionTestCase for unit testing
     """
@@ -113,6 +116,16 @@ class TransactionTestCase(_TransactionTestCase):
 
         # Re-populate status choices after database truncation by TransactionTestCase
         populate_status_choices(apps, None)
+
+        # Create the test user and assign permissions
+        self.user = User.objects.create_user(username="testuser")
+        self.add_permissions(*self.user_permissions)
+
+        # Initialize the test client
+        self.client = Client()
+
+        # Force login explicitly with the first-available backend
+        self.client.force_login(self.user)
 
 
 class CeleryTestCase(TransactionTestCase):
